Add Api module for the command call of ElasticDL client.
